fix: 32bits build

+++ src/catfs/dir.rs
@@ -4,13 +4,14 @@ use std::fs;
 use std::io;
 use std::os::unix::io::RawFd;
 use std::path::Path;
+use std::os::raw::c_int;
 
 use catfs::error;
 use catfs::rlibc;
 
 pub struct Handle {
     dh: *mut libc::DIR,
-    offset: i64,
+    offset: c_int,
     entry: rlibc::Dirent,
     entry_valid: bool,
 }
@@ -60,7 +61,7 @@ impl Handle {
         });
     }
 
-    pub fn seekdir(&mut self, offset: i64) {
+    pub fn seekdir(&mut self, offset: c_int) {
         if offset != self.offset {
             debug!(
                 "seeking {} to {}",
